+++ lib/dialogs/scroll_picker_dialog.dart
@@ -23,7 +23,7 @@ class ScrollPickerDialog<T> extends StatefulWidget
     this.buttonTextColor,
     this.maxLongSide,
     this.maxShortSide,
-    this.showDivider = true,
+    this.showDivider = false,
     this.confirmText,
     this.cancelText,
     this.cancelButtonVisible = true,